rest_apis_project/requirements.txt
fastapi==0.111.0
bcrypt==4.1.3
pydantic==2.7.1
pydantic-settings==2.2.1
pyjwt==2.8.0
pymongo==4.7.0
